dom/performance/tests/test_performance_timing_json.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1375829
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1375829</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 1375829 **/
  var json = performance.timing.toJSON();

  // Ensure it doesn't have any attributes that performance.timing doesn't have
  for (let key of Object.keys(json)) {
    ok(key in performance.timing, key + " should be a property of performance.timing");
  }

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1375829">Mozilla Bug 1375829</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
